+++ test/e2e.test.ts
@@ -1,4 +1,5 @@
 import { chromium, Page, Browser } from "playwright"
+import { CODE_SERVER_ADDRESS } from "./constants"
 
 let browser: Browser
 let page: Page
@@ -17,7 +18,7 @@ afterEach(async () => {
 })
 
 it("should see the login page", async () => {
-  await page.goto(process.env)
+  await page.goto(CODE_SERVER_ADDRESS)
   // It should send us to the login page
   expect(await page.title()).toBe("code-server login")
 })
